refactor(parser): use nom_locate in parser
nom_locate provides a type (LocatedSpan) that tracks the location within the
input data as it is parsed. This will allow us to track the location of
transactions and postings within a journal, which is useful right now for
improving the account/commodity checking errors (which report the line number
and content of the error). It may also be useful when we get to supporting
output formatting like %b (start line of a transaction)

Although this uses nom_locate throughout the parser, it doesn't yet take
advantage of any of the actual location-aware features. It does, however,
compile and pass tests.